Corrected the Code Separation
The code is now more appropriately separated into header and source
files.

I still need to refractor the main method into a utility function, but
since this is a utility and not a library, it's not inherently
necessary.

I should definitely still do it though in case I wish to expand.